pyproject.toml
[project]
name = "linkedin-mutual-connections"
version = "0.1.0"
description = "LLM-driven agent that extracts mutual LinkedIn connections via browser-use + Gemini"
requires-python = ">=3.11"
dependencies = [
    "browser-use==0.12.0",
    "playwright==1.58.0",
    "google-genai==1.60.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.41",
    "google-cloud-firestore>=2.19",
    "google-cloud-pubsub>=2.0",
    "pydantic-settings>=2.0",
]

[project.scripts]
save-cookies = "save_cookies:main_sync"
mutual-connections = "mutual_connections:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
